requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
django-filter>=21.1,<22.0
psycopg2>=2.9,<3.0  
python-dotenv>=0.19,<1.0 
djangorestframework-simplejwt>=5.0,<6.0  
mysqlclient>=2.1,<3.0